Fix: Output namespace in inner <svg> element.
